fix search bug, add update known faces module

+++ server/drop_database.py
@@ -9,8 +9,10 @@ db = cl["miris"]
 col = db["face"]
 col.drop()
 
-shutil.rmtree("images/customer")
-shutil.rmtree("data/arcface_feats")
+if os.path.exists("images/customer"):
+    shutil.rmtree("images/customer")
+if os.path.exists("data/arcface_feats"):
+    shutil.rmtree("data/arcface_feats")
 
 os.mkdir("images/customer")
 # os.mkdir("data/arcface_feats")
